+++ .github/type-coverage.json
@@ -2,8 +2,8 @@
   "succeeded": true,
   "atLeast": 90,
   "atLeastFailed": false,
-  "correctCount": 4625,
-  "percent": 93.02,
-  "percentString": "93.02",
-  "totalCount": 4972
+  "correctCount": 4809,
+  "percent": 93.12,
+  "percentString": "93.12",
+  "totalCount": 5164
 }
