+++ src/routes/users/getSelf.ts
@@ -1,7 +1,7 @@
 import { Router } from "express"
 import { authenticateToken } from "../../middleware/auth/jsonWebTokens"
 import CONTROLLERS from "../../controllers"
-import RBAC from "../../middleware/auth/roleBasedAccessControl"
+import RBAC from "../../middleware/auth/rbac/roleBasedAccessControl"
 
 const router = Router()
