+++ src/routes/users/getUser.ts
@@ -1,6 +1,6 @@
 import { Router } from "express"
 import { authenticateToken } from "../../middleware/auth/jsonWebTokens"
-import RBAC from "../../middleware/auth/roleBasedAccessControl"
+import RBAC from "../../middleware/auth/rbac/roleBasedAccessControl"
 import CONTROLLERS from "../../controllers"
 
 const router = Router()
