+++ src/routes/users/updateUser.ts
@@ -1,10 +1,10 @@
 import { Router } from "express"
 import { authenticateToken } from "../../middleware/auth/jsonWebTokens"
-import RBAC from "../../middleware/auth/roleBasedAccessControl"
+import RBAC from "../../middleware/auth/rbac/roleBasedAccessControl"
 import CONTROLLERS from "../../controllers"
 
 const router = Router()
 
-router.patch("/:id", authenticateToken, RBAC.getTargetUser, RBAC.checkCanUpdateUsers, CONTROLLERS.USER.updateUser)
+router.patch("/:id", authenticateToken, RBAC.getTargetUser, RBAC.checkCanUpdateUser, CONTROLLERS.USER.updateUser)
 
 export default router
